Move renderAction into ContentHoverWidget
